Fix Windows cross-compilation CI issues
- Exclude arena_win32.c from clang-tidy (no windows.h on Linux host)
- Fix Conan profile: use tools.cmake.cmaketoolchain:user_toolchain
  instead of nonexistent tools.cmake:cmake_toolchain_file conf
- Fix bugprone-implicit-widening-of-multiplication-result in
  WIN32_COMMIT_GRANULARITY by casting to size_t

+++ engine/src/arena_win32.c
@@ -8,7 +8,7 @@
 #include <string.h>
 #include <windows.h>
 
-#define WIN32_COMMIT_GRANULARITY (64U * 1024U) /* 64 KiB — Windows allocation granularity */
+#define WIN32_COMMIT_GRANULARITY ((size_t)64U * 1024U) /* 64 KiB — Windows allocation granularity */
 
 bool arena_init(ErrorState *err, Arena *arena)
 {
